+++ src/Pluto.jl
@@ -26,13 +26,14 @@ include("./analysis/ExpressionExplorer.jl")
 include("./analysis/ReactiveNode.jl")
 
 include("./notebook/Cell.jl")
+include("./analysis/Topology.jl")
 include("./notebook/Notebook.jl")
 include("./webserver/Session.jl")
 include("./webserver/PutUpdates.jl")
 
 include("./analysis/Errors.jl")
 include("./analysis/Parse.jl")
-include("./analysis/Topology.jl")
+include("./analysis/TopologicalOrder.jl")
 
 include("./evaluation/WorkspaceManager.jl")
 include("./evaluation/Update.jl")
