+++ src/analysis/Errors.jl
@@ -8,8 +8,8 @@ struct CyclicReferenceError <: ReactivityError
 end
 
 function CyclicReferenceError(topology::NotebookTopology, cycle::Cell...)
-	referenced_during_cycle = union((topology[c].references for c in cycle)...)
-	assigned_during_cycle = union((topology[c].definitions ∪ topology[c].funcdefs_without_signatures for c in cycle)...)
+	referenced_during_cycle = union((topology.nodes[c].references for c in cycle)...)
+	assigned_during_cycle = union((topology.nodes[c].definitions ∪ topology.nodes[c].funcdefs_without_signatures for c in cycle)...)
 	
 	CyclicReferenceError(referenced_during_cycle ∩ assigned_during_cycle)
 end
@@ -20,7 +20,7 @@ end
 
 function MultipleDefinitionsError(topology::NotebookTopology, cell::Cell, all_definers)
 	competitors = setdiff(all_definers, [cell])
-	defs(c) = topology[c].funcdefs_without_signatures ∪ topology[c].definitions
+	defs(c) = topology.nodes[c].funcdefs_without_signatures ∪ topology.nodes[c].definitions
 	MultipleDefinitionsError(
 		union((defs(cell) ∩ defs(c) for c in competitors)...)
 	)
