+++ test/frontend/README.md
@@ -6,23 +6,23 @@
 
 ## Run Pluto.jl server
 
-`PLUTO_PORT=1235 julia --project=/path/to/PlutoDev -e "import Pluto; Pluto.run($PLUTO_PORT)"`
+`PLUTO_PORT=2345 julia --project=/path/to/PlutoDev -e "import Pluto; Pluto.run(port=$PLUTO_PORT, require_secret_for_access=false, require_secret_for_open_links=false, launch_browser=false)"`
 
 ## Run tests
 
-`PLUTO_PORT=1235 npm run test`
+`PLUTO_PORT=2345 npm run test`
 
 ## View the browser in action
 
 Add `HEADLESS=false` when running the test command.
 
-`HEADLESS=false PLUTO_PORT=1235 npm run test`
+`clear && HEADLESS=false PLUTO_PORT=2345 npm run test`
 
 ## Run a particular suite of tests
 
 Add `-- -t=name of the suite` to the end of the test command.
 
-`PLUTO_PORT=1235 npm run test -- -t=PlutoAutocomplete`
+`clear && PLUTO_PORT=2345 npm run test -- -t=PlutoAutocomplete`
 
 ## To make a test fail on a case that does not crash Pluto
